backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-unfold>=0.21.0
django-parler>=2.3
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
django-parler-rest>=2.2
django-filter>=23.0